Private collections on Collections lists
Implement the concept of proper private collections in the CollectionsList and
Collection.

When creating a collection, it has a $public attribute that defaults to true.
When creating a collections list, it can be with private items, or false.
  Take into account that watchlist is private by default.
In special collections, checking if the user logged in is the owner of the
current thing we are going to show, pass privacy options to the CollectionsList
model.

This fixes going not logged in to SpecialCollections/User and seeing the
watchlist information (number of articles).